# web/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
    "/subjects/search": {
      "post": {
        "summary": "Search Subjects",
        "description": "Search subjects using Elasticsearch.",
        "operationId": "search_subjects_subjects_search_post",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SubjectsIndexQuery" } } },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SubjectSearchResult" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/characters/search": {
      "post": {
        "summary": "Search Characters",
        "description": "Search characters using Elasticsearch.",
        "operationId": "search_characters_characters_search_post",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CharactersIndexQuery" } } },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CharacterSearchResult" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/people/search": {
      "post": {
        "summary": "Search People",
        "description": "Search people using Elasticsearch.",
        "operationId": "search_people_people_search_post",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PersonsIndexQuery" } } },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PersonSearchResult" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/episodes/search": {
      "post": {
        "summary": "Search Episodes",
        "description": "Search episodes using Elasticsearch.",
        "operationId": "search_episodes_episodes_search_post",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EpisodesIndexQuery" } } },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EpisodeSearchResult" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/subjects/{subject_id}": {
      "get": {
        "summary": "Get Subject",
        "operationId": "get_subject_subjects__subject_id__get",
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Subject Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Subject" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/subjects/multiple": {
      "get": {
        "summary": "Get Subjects Multiple",
        "operationId": "get_subjects_multiple_subjects_multiple_get",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Ids" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Subject" },
                  "title": "Response Get Subjects Multiple Subjects Multiple Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/subjects/{subject_id}/episodes": {
      "get": {
        "summary": "Get Subject Episodes",
        "operationId": "get_subject_episodes_subjects__subject_id__episodes_get",
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Subject Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Episode" },
                  "title": "Response Get Subject Episodes Subjects  Subject Id  Episodes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/subjects/{subject_id}/edges": {
      "get": {
        "summary": "Get Subject Edges",
        "operationId": "get_subject_edges_subjects__subject_id__edges_get",
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Subject Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Subgraph" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/characters/{character_id}": {
      "get": {
        "summary": "Get Character",
        "operationId": "get_character_characters__character_id__get",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Character Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Character" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/characters/multiple": {
      "get": {
        "summary": "Get Characters Multiple",
        "operationId": "get_characters_multiple_characters_multiple_get",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Ids" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Character" },
                  "title": "Response Get Characters Multiple Characters Multiple Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/people/{person_id}": {
      "get": {
        "summary": "Get Person",
        "operationId": "get_person_people__person_id__get",
        "parameters": [
          { "name": "person_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Person Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Person" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/people/multiple": {
      "get": {
        "summary": "Get People Multiple",
        "operationId": "get_people_multiple_people_multiple_get",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Ids" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Person" },
                  "title": "Response Get People Multiple People Multiple Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnimeStuff": {
        "type": "integer",
        "enum": [
          1,
          74,
          2,
          72,
          3,
          4,
          89,
          5,
          91,
          6,
          7,
          8,
          9,
          10,
          11,
          71,
          13,
          16,
          19,
          14,
          15,
          90,
          70,
          77,
          17,
          69,
          86,
          18,
          20,
          21,
          92,
          22,
          67,
          73,
          82,
          65,
          85,
          55,
          24,
          25,
          26,
          27,
          75,
          37,
          28,
          29,
          30,
          31,
          32,
          33,
          34,
          35,
          36,
          38,
          39,
          40,
          41,
          42,
          63,
          43,
          84,
          44,
          45,
          46,
          88,
          47,
          48,
          49,
          50,
          51,
          58,
          59,
          54,
          52,
          53,
          23,
          87,
          80,
          56,
          83,
          57,
          60,
          61,
          62,
          64,
          76,
          66,
          81
        ],
        "title": "AnimeStuff"
      },
      "BookStaff": {
        "type": "integer",
        "enum": [2007, 2001, 2002, 2003, 2010, 2004, 2005, 2006, 2008, 2009, 2011, 2012, 2013],
        "title": "BookStaff"
      },
      "Character": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "role": { "$ref": "#/components/schemas/CharacterRole" },
          "name": { "type": "string", "title": "Name" },
          "infobox": { "type": "string", "title": "Infobox" },
          "summary": { "type": "string", "title": "Summary" },
          "comments": { "type": "integer", "title": "Comments" },
          "collects": { "type": "integer", "title": "Collects" }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["id", "role", "name", "infobox", "summary", "comments", "collects"],
        "title": "Character",
        "description": "Character model."
      },
      "CharacterRole": {
        "type": "integer",
        "enum": [1, 2, 3, 4],
        "title": "CharacterRole",
        "description": "Character role types in Bangumi."
      },
      "CharacterSearchResult": {
        "properties": {
          "items": { "items": { "$ref": "#/components/schemas/Character" }, "type": "array", "title": "Items" },
          "total": { "type": "integer", "title": "Total" },
          "query": { "type": "string", "title": "Query" },
          "limit": { "type": "integer", "title": "Limit" },
          "offset": { "type": "integer", "title": "Offset" },
          "has_more": { "type": "boolean", "title": "Has More" }
        },
        "type": "object",
        "required": ["items", "total", "query", "limit", "offset", "has_more"],
        "title": "CharacterSearchResult",
        "description": "Result model for character search operations."
      },
      "CharactersIndexQuery": {
        "properties": {
          "query": { "type": "string", "title": "Query" },
          "limit": { "type": "integer", "title": "Limit", "default": 20 },
          "offset": { "type": "integer", "title": "Offset", "default": 0 },
          "role": { "anyOf": [{ "type": "integer" }, { "type": "null" }], "title": "Role" }
        },
        "type": "object",
        "required": ["query"],
        "title": "CharactersIndexQuery"
      },
      "Episode": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "name_cn": { "type": "string", "title": "Name Cn" },
          "description": { "type": "string", "title": "Description" },
          "airdate": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Airdate" },
          "disc": { "type": "integer", "title": "Disc" },
          "duration": { "type": "string", "title": "Duration" },
          "subject_id": { "type": "integer", "title": "Subject Id" },
          "sort": { "anyOf": [{ "type": "integer" }, { "type": "number" }], "title": "Sort" },
          "type": { "$ref": "#/components/schemas/EpisodeType" }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["id", "name", "name_cn", "description", "disc", "duration", "subject_id", "sort", "type"],
        "title": "Episode",
        "description": "Episode model."
      },
      "EpisodeSearchResult": {
        "properties": {
          "items": { "items": { "$ref": "#/components/schemas/Episode" }, "type": "array", "title": "Items" },
          "total": { "type": "integer", "title": "Total" },
          "query": { "type": "string", "title": "Query" },
          "limit": { "type": "integer", "title": "Limit" },
          "offset": { "type": "integer", "title": "Offset" },
          "has_more": { "type": "boolean", "title": "Has More" }
        },
        "type": "object",
        "required": ["items", "total", "query", "limit", "offset", "has_more"],
        "title": "EpisodeSearchResult",
        "description": "Result model for episode search operations."
      },
      "EpisodeType": {
        "type": "integer",
        "enum": [0, 1, 2, 3, 4, 5, 6],
        "title": "EpisodeType",
        "description": "Episode types in Bangumi."
      },
      "EpisodesIndexQuery": {
        "properties": {
          "query": { "type": "string", "title": "Query" },
          "limit": { "type": "integer", "title": "Limit", "default": 20 },
          "offset": { "type": "integer", "title": "Offset", "default": 0 },
          "type": { "anyOf": [{ "type": "integer" }, { "type": "null" }], "title": "Type" },
          "subject_id": { "anyOf": [{ "type": "integer" }, { "type": "null" }], "title": "Subject Id" }
        },
        "type": "object",
        "required": ["query"],
        "title": "EpisodesIndexQuery"
      },
      "Favorite": {
        "properties": {
          "wish": { "type": "integer", "title": "Wish" },
          "done": { "type": "integer", "title": "Done" },
          "doing": { "type": "integer", "title": "Doing" },
          "on_hold": { "type": "integer", "title": "On Hold" },
          "dropped": { "type": "integer", "title": "Dropped" }
        },
        "type": "object",
        "required": ["wish", "done", "doing", "on_hold", "dropped"],
        "title": "Favorite",
        "description": "Favorite statistics."
      },
      "GameStaff": {
        "type": "integer",
        "enum": [
          1001,
          1002,
          1003,
          1015,
          1016,
          1032,
          1028,
          1026,
          1004,
          1027,
          1031,
          1013,
          1008,
          1029,
          1005,
          1023,
          1024,
          1025,
          1030,
          1006,
          1021,
          1014,
          1017,
          1020,
          1018,
          1019,
          1007,
          1009,
          1010,
          1011,
          1012,
          1022
        ],
        "title": "GameStaff"
      },
      "GraphEdge": {
        "properties": {
          "to_subject": { "anyOf": [{ "$ref": "#/components/schemas/Subject" }, { "type": "null" }] },
          "to_character": { "anyOf": [{ "$ref": "#/components/schemas/Character" }, { "type": "null" }] },
          "to_person": { "anyOf": [{ "$ref": "#/components/schemas/Person" }, { "type": "null" }] },
          "s2s_relation_type": {
            "anyOf": [{ "$ref": "#/components/schemas/SubjectRelationType" }, { "type": "null" }]
          },
          "p2c_summary": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "P2C Summary" },
          "s2p_position": {
            "anyOf": [
              { "$ref": "#/components/schemas/AnimeStuff" },
              { "$ref": "#/components/schemas/BookStaff" },
              { "$ref": "#/components/schemas/GameStaff" },
              { "$ref": "#/components/schemas/MusicStaff" },
              { "$ref": "#/components/schemas/RealStaff" },
              { "type": "null" }
            ],
            "title": "S2P Position"
          },
          "s2c_type": { "anyOf": [{ "$ref": "#/components/schemas/SubjectCharacterType" }, { "type": "null" }] }
        },
        "type": "object",
        "title": "GraphEdge"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": { "items": { "$ref": "#/components/schemas/ValidationError" }, "type": "array", "title": "Detail" }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MusicStaff": {
        "type": "integer",
        "enum": [3001, 3002, 3004, 3003, 3006, 3008, 3014, 3015, 3007, 3013, 3012, 3009, 3005, 3010, 3011],
        "title": "MusicStaff"
      },
      "Person": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "type": { "$ref": "#/components/schemas/PersonType" },
          "career": { "items": { "type": "string" }, "type": "array", "title": "Career" },
          "infobox": { "type": "string", "title": "Infobox" },
          "summary": { "type": "string", "title": "Summary" },
          "comments": { "type": "integer", "title": "Comments" },
          "collects": { "type": "integer", "title": "Collects" }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["id", "name", "type", "infobox", "summary", "comments", "collects"],
        "title": "Person",
        "description": "Person model (individual, company, association)."
      },
      "PersonSearchResult": {
        "properties": {
          "items": { "items": { "$ref": "#/components/schemas/Person" }, "type": "array", "title": "Items" },
          "total": { "type": "integer", "title": "Total" },
          "query": { "type": "string", "title": "Query" },
          "limit": { "type": "integer", "title": "Limit" },
          "offset": { "type": "integer", "title": "Offset" },
          "has_more": { "type": "boolean", "title": "Has More" }
        },
        "type": "object",
        "required": ["items", "total", "query", "limit", "offset", "has_more"],
        "title": "PersonSearchResult",
        "description": "Result model for person search operations."
      },
      "PersonType": {
        "type": "integer",
        "enum": [0, 1, 2, 3],
        "title": "PersonType",
        "description": "Person types in Bangumi."
      },
      "PersonsIndexQuery": {
        "properties": {
          "query": { "type": "string", "title": "Query" },
          "limit": { "type": "integer", "title": "Limit", "default": 20 },
          "offset": { "type": "integer", "title": "Offset", "default": 0 },
          "type": { "anyOf": [{ "type": "integer" }, { "type": "null" }], "title": "Type" },
          "career": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Career" }
        },
        "type": "object",
        "required": ["query"],
        "title": "PersonsIndexQuery"
      },
...
      "ScoreDetails": {
        "properties": {
          "1": { "type": "integer", "title": "1", "default": 0 },
          "2": { "type": "integer", "title": "2", "default": 0 },
          "3": { "type": "integer", "title": "3", "default": 0 },
          "4": { "type": "integer", "title": "4", "default": 0 },
          "5": { "type": "integer", "title": "5", "default": 0 },
          "6": { "type": "integer", "title": "6", "default": 0 },
          "7": { "type": "integer", "title": "7", "default": 0 },
          "8": { "type": "integer", "title": "8", "default": 0 },
          "9": { "type": "integer", "title": "9", "default": 0 },
          "10": { "type": "integer", "title": "10", "default": 0 }
        },
        "type": "object",
        "title": "ScoreDetails",
        "description": "Score distribution details."
      },
      "Subgraph": {
        "properties": {
          "from_subject": { "anyOf": [{ "$ref": "#/components/schemas/Subject" }, { "type": "null" }] },
          "from_character": { "anyOf": [{ "$ref": "#/components/schemas/Character" }, { "type": "null" }] },
          "from_person": { "anyOf": [{ "$ref": "#/components/schemas/Person" }, { "type": "null" }] },
          "edges": { "items": { "$ref": "#/components/schemas/GraphEdge" }, "type": "array", "title": "Edges" }
        },
        "type": "object",
        "required": ["edges"],
        "title": "Subgraph",
        "description": "A subgraph of vertices/edges around to a \"from\" vertex"
      },
      "Subject": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "type": { "$ref": "#/components/schemas/SubjectType" },
          "name": { "type": "string", "title": "Name" },
          "name_cn": { "type": "string", "title": "Name Cn" },
          "infobox": { "type": "string", "title": "Infobox" },
          "platform": { "type": "integer", "title": "Platform" },
          "summary": { "type": "string", "title": "Summary" },
          "nsfw": { "type": "boolean", "title": "Nsfw" },
          "tags": { "items": { "$ref": "#/components/schemas/Tag" }, "type": "array", "title": "Tags" },
          "score": { "type": "number", "title": "Score" },
          "score_details": { "anyOf": [{ "$ref": "#/components/schemas/ScoreDetails" }, { "type": "null" }] },
          "rank": { "type": "integer", "title": "Rank" },
          "date": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Date" },
          "favorite": { "$ref": "#/components/schemas/Favorite" },
          "series": { "type": "boolean", "title": "Series" },
          "meta_tags": {
            "anyOf": [{ "items": { "type": "string" }, "type": "array" }, { "type": "null" }],
            "title": "Meta Tags"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "name",
          "name_cn",
          "infobox",
          "platform",
          "summary",
          "nsfw",
          "tags",
          "score",
          "rank",
          "date",
          "favorite",
          "series"
        ],
        "title": "Subject",
        "description": "Subject model (anime, book, game, etc.)."
      },
      "SubjectCharacterType": {
        "type": "integer",
        "enum": [1, 2, 3, 4, 5],
        "title": "SubjectCharacterType",
        "description": "Character types in subject-character relationships."
      },
      "SubjectRelationType": {
        "type": "integer",
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11,
          12,
          14,
          99,
          1002,
          1003,
          1004,
          1005,
          1006,
          1007,
          1008,
          1010,
          1011,
          1012,
          1013,
          1014,
          1015,
          1099,
          3001,
          3002,
          3003,
          3004,
          3005,
          3006,
          3007,
          3099,
          4002,
          4003,
          4006,
          4007,
          4008,
          4009,
          4010,
          4012,
          4014,
          4015,
          4016,
          4017,
          4018,
          4019,
          4099
        ],
        "title": "SubjectRelationType",
        "description": "Relation types between subjects."
      },
      "SubjectSearchResult": {
        "properties": {
          "items": { "items": { "$ref": "#/components/schemas/Subject" }, "type": "array", "title": "Items" },
          "total": { "type": "integer", "title": "Total" },
          "query": { "type": "string", "title": "Query" },
          "limit": { "type": "integer", "title": "Limit" },
          "offset": { "type": "integer", "title": "Offset" },
          "has_more": { "type": "boolean", "title": "Has More" }
        },
        "type": "object",
        "required": ["items", "total", "query", "limit", "offset", "has_more"],
        "title": "SubjectSearchResult",
        "description": "Result model for subject search operations."
      },
      "SubjectType": {
        "type": "integer",
        "enum": [1, 2, 3, 4, 6],
        "title": "SubjectType",
        "description": "Subject types in Bangumi."
      },
      "SubjectsIndexQuery": {
        "properties": {
          "query": { "type": "string", "title": "Query" },
          "limit": { "type": "integer", "title": "Limit", "default": 20 },
          "offset": { "type": "integer", "title": "Offset", "default": 0 },
          "type": { "anyOf": [{ "type": "integer" }, { "type": "null" }], "title": "Type" },
          "subject_type": { "anyOf": [{ "type": "integer" }, { "type": "null" }], "title": "Subject Type" },
          "nsfw": { "anyOf": [{ "type": "boolean" }, { "type": "null" }], "title": "Nsfw" }
        },
        "type": "object",
        "required": ["query"],
        "title": "SubjectsIndexQuery"
      },
      "Tag": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["name", "count"],
        "title": "Tag",
        "description": "Tag model for subjects."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}